% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{jqpd_sd}
\alias{jqpd_sd}
\title{Calculates the standard-deviation of a Johnson Quantile-Parameterised
Distribution.}
\usage{
jqpd_sd(params)
}
\arguments{
\item{params}{jqpd object created using \code{jqpd()}}
}
\description{
Calculates the standard-deviation of a Johnson Quantile-Parameterised
Distribution.
}
