% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{jqpd_var}
\alias{jqpd_var}
\title{Calculates the variance of a Johnson Quantile-Parameterised Distribution.}
\usage{
jqpd_var(params)
}
\arguments{
\item{params}{jqpd object created using \code{jqpd()}}
}
\description{
Calculates the variance of a Johnson Quantile-Parameterised Distribution.
}
