% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_jqpd}
\alias{plot_jqpd}
\title{Plots the density, cumulative distribution function, quantile function and a
set of 1000 random samples for a Johnson Quantile-Parameterised Distribution.}
\usage{
plot_jqpd(params)
}
\arguments{
\item{params}{jqpd object created using \code{jqpd()}}
}
\description{
Plots the density, cumulative distribution function, quantile function and a
set of 1000 random samples for a Johnson Quantile-Parameterised Distribution.
}
