% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjqpd.R
\name{rjqpd}
\alias{rjqpd}
\title{Generate random samples from a jqpd distribution object}
\usage{
rjqpd(n = 1, params)
}
\arguments{
\item{n}{number of observations (default is 1)}

\item{params}{jqpd object created using \code{jqpd()}}
}
\value{
A numeric vector of n random samples from the input distribution
}
\description{
Generate random samples from a jqpd distribution object
}
\examples{
x <- c(0.32, 0.40, 0.60)
params <- jqpd(x, lower = 0, upper = 1, alpha = 0.1)
samples <- rjqpd(n = 1000, params)
}
