% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{jqpd_kurtosis}
\alias{jqpd_kurtosis}
\title{Calculates the kurtosis of a Johnson Quantile-Parameterised
Distribution.}
\usage{
jqpd_kurtosis(params)
}
\arguments{
\item{params}{jqpd object created using \code{jqpd()}}
}
\value{
The kurtosis of the distribution as a length one numeric vector.
}
\description{
Calculates the kurtosis of a Johnson Quantile-Parameterised
Distribution.
}
