% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{jqpd_mean}
\alias{jqpd_mean}
\title{Calculates the mean of a Johnson Quantile-Parameterised Distribution.}
\usage{
jqpd_mean(params)
}
\arguments{
\item{params}{jqpd object created using \code{jqpd()}}
}
\value{
The mean of the distribution as a length one numeric vector.
}
\description{
Calculates the mean of a Johnson Quantile-Parameterised Distribution.
}
