% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{jqpd_skewness}
\alias{jqpd_skewness}
\title{Calculates the skewness of a Johnson Quantile-Parameterised
Distribution.}
\usage{
jqpd_skewness(params)
}
\arguments{
\item{params}{jqpd object created using \code{jqpd()}}
}
\description{
Calculates the skewness of a Johnson Quantile-Parameterised
Distribution.
}
